backend/requirements.txt
# Core Dependencies
flask==2.3.3
numpy>=1.21.0
Pillow>=8.0.0
python-dotenv>=1.0.0

# AI Core Dependencies (DINOv2 + YOLOv8)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.25.0
ultralytics>=8.0.0

# Vector Database - FAISS (CPU version)
faiss-cpu>=1.7.0

# Web Scraping Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Discord Bot Dependencies
discord.py-self>=2.0.0
aiohttp>=3.8.0
